massive overhaul of main form - attempted to divide and conquer lots of form elements here into 7 partial classes - mostly for organizing, lays the groundwork for future refactoring - makes it easier to mentally digest the important bits separate from the glue
